backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Balashov Igor",email = "igor2233balashov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
